Refactor the edit-schedule pack
And also introduce a wcif-utils to share between edit-events and edit-schedule
