+++ WcaOnRails/app/assets/stylesheets/edit_schedule.scss
@@ -25,6 +25,7 @@
           margin-bottom: 10px;
         }
       }
+      margin-bottom: 15px;
     }
   }
