+++ WcaOnRails/app/javascript/edit-events/modals/AdvancementCondition.jsx
@@ -1,7 +1,8 @@
 import React from 'react'
 
 import AttemptResultInput from './AttemptResultInput'
-import { attemptResultToString, roundIdToString, matchResult } from './utils'
+import { attemptResultToString, matchResult } from './utils'
+import { roundIdToString } from 'wca/wcif-utils'
 
 const MIN_ADVANCE_PERCENT = 1;
 const MAX_ADVANCE_PERCENT = 75;
