+++ WcaOnRails/app/javascript/edit-events/modals/utils.js
@@ -109,43 +109,3 @@ export function centisecondsToString(centiseconds, { short } = {}) {
 
   return str.trim();
 }
-
-export function roundIdToString(roundId) {
-  let { eventId, roundNumber } = parseActivityCode(roundId);
-  let event = events.byId[eventId];
-  return `${event.name}, Round ${roundNumber}`;
-}
-
-// Copied from https://github.com/jfly/tnoodle/blob/c2b529e6292469c23f33b1d73839e22f041443e0/tnoodle-ui/src/WcaCompetitionJson.js#L52
-export function parseActivityCode(activityCode) {
-  let eventId, roundNumber, group, attempt;
-  let parts = activityCode.split("-");
-  eventId = parts.shift();
-
-  parts.forEach(part => {
-    let firstLetter = part[0];
-    let rest = part.substring(1);
-    if(firstLetter === "r") {
-      roundNumber = parseInt(rest, 10);
-    } else if(firstLetter === "g") {
-      group = rest;
-    } else if(firstLetter === "a") {
-      attempt = rest;
-    } else {
-      throw new Error(`Unrecognized activity code part: ${part} of ${activityCode}`);
-    }
-  });
-  return { eventId, roundNumber, group, attempt };
-}
-
-export function buildActivityCode(activity) {
-  let activityCode = activity.eventId;
-  if(activity.roundNumber) {
-    activityCode += "-r" + activity.roundNumber;
-  }
-  if(activity.group) {
-    activityCode += "-g" + activity.group;
-  }
-
-  return activityCode;
-}
